Refactor: Reorganize Card component and replace divs with sections
