+++ src/components/Cards/components/BackCard.tsx
@@ -1,0 +1,9 @@
+export default function BackCard() {
+    return (
+        <div className="bg-[url('assets/images/bg-card-back.png')] bg-cover w-72 h-40 md:w-96 md:h-52 rounded-lg p-6 text-white text-xs md:text-sm tracking-widest lg:col-start-2 lg:col-end-5 lg:static absolute top-5 right-5 sm:top-9 sm:right-6">
+            <p className="h-full flex justify-end items-center pr-2.5 md:pr-5">
+                000
+            </p>
+        </div>
+    )
+}
